feat(inversify-code-examples): rename examples
