+++ packages/docs/tools/inversify-code-examples/src/examples/bindingScopeRequest.ts
