+++ packages/docs/tools/inversify-code-examples/src/examples/bindingScopeSingleton.int.spec.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from '@jest/globals';
 
-import { isSameKatana } from './bindingInSyntaxApiInSingletonScope';
+import { isSameKatana } from './bindingScopeSingleton';
 
 describe('BindingInSyntax API (inSingletonScope)', () => {
   it('should provide same Katana', () => {
