+++ packages/docs/tools/inversify-code-examples/src/examples/bindingScopeSingleton.ts
