+++ packages/docs/tools/inversify-code-examples/src/examples/bindingScopeTransient.int.spec.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from '@jest/globals';
 
-import { isSameKatana } from './bindingInSyntaxApiInTransientScope';
+import { isSameKatana } from './bindingScopeTransient';
 
 describe('BindingInSyntax API (inTransientScope)', () => {
   it('should provide same Katana', () => {
