+++ packages/docs/tools/inversify-code-examples/src/examples/bindingScopeTransient.ts
